Support/requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cymem==2.0.6
DAWG-Python==0.7.2
Django==4.0.4
djangorestframework==3.13.1
docopt==0.6.2
frozenlist==1.3.0
idna==3.3
install==1.3.5
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
multidict==6.0.2
murmurhash==1.0.7
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
regex==2022.4.24
requests==2.27.1
six==1.16.0
smart-open==5.2.1
spacy==3.2.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
spacy-russian-tokenizer==0.1.1
sqlparse==0.4.2
srsly==2.4.3
thinc==8.0.15
torch==1.11.0
tqdm==4.64.0
transliterate==1.10.2
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
yarl==1.7.2
